+++ deepdocforgery/__init__.py
@@ -3,4 +3,4 @@
 from deepdocforgery.model import DeepDocForgeryModel, DeepDocForgeryOutput
 
 __all__ = ["DeepDocForgeryModel", "DeepDocForgeryOutput"]
-__version__ = "0.5.1"
+__version__ = "0.5.2"